separate out Plugin.GHCSpecter.Util library
