+++ plugin/src/Plugin/GHCSpecter/Util.hs
@@ -1,0 +1,130 @@
+module Plugin.GHCSpecter.Util
+  ( -- * Utilities
+    getTopSortedModules,
+    extractModuleGraphInfo,
+    getModuleNameFromPipeState,
+    mkModuleNameMap,
+    formatName,
+    formatImportedNames,
+  )
+where
+
+import Data.Char (isAlpha)
+import Data.IntMap (IntMap)
+import Data.IntMap qualified as IM
+import Data.List qualified as L
+import Data.Map (Map)
+import Data.Map qualified as M
+import Data.Maybe (mapMaybe)
+import Data.Text qualified as T
+import Data.Tuple (swap)
+import GHC.Data.Graph.Directed qualified as G
+import GHC.Driver.Make
+  ( moduleGraphNodes,
+    topSortModuleGraph,
+  )
+import GHC.Driver.Pipeline
+  ( PipeState (iface),
+  )
+import GHC.Driver.Session (DynFlags)
+import GHC.Plugins
+  ( ModSummary,
+    Name,
+    localiseName,
+    showSDoc,
+  )
+import GHC.Types.Name.Reader
+  ( GlobalRdrElt (..),
+    GreName (..),
+    ImpDeclSpec (..),
+    ImportSpec (..),
+  )
+import GHC.Unit.Module.Graph
+  ( ModuleGraph,
+    ModuleGraphNode (..),
+    mgModSummaries',
+  )
+import GHC.Unit.Module.ModIface (ModIface_ (mi_module))
+import GHC.Unit.Module.ModSummary
+  ( ExtendedModSummary (..),
+    ModSummary (..),
+  )
+import GHC.Unit.Module.Name (moduleNameString)
+import GHC.Unit.Types (GenModule (moduleName))
+import GHC.Utils.Outputable (Outputable (ppr))
+import GHCSpecter.Channel.Common.Types (type ModuleName)
+import GHCSpecter.Channel.Outbound.Types (ModuleGraphInfo (..))
+
+getModuleNameText :: ModSummary -> T.Text
+getModuleNameText = T.pack . moduleNameString . moduleName . ms_mod
+
+gnode2ModSummary :: ModuleGraphNode -> Maybe ModSummary
+gnode2ModSummary InstantiationNode {} = Nothing
+gnode2ModSummary (ModuleNode emod) = Just (emsModSummary emod)
+
+-- temporary function. ignore hs-boot cycles and InstantiatedUnit for now.
+getTopSortedModules :: ModuleGraph -> [ModuleName]
+getTopSortedModules =
+  mapMaybe (fmap getModuleNameText . gnode2ModSummary)
+    . concatMap G.flattenSCC
+    . flip (topSortModuleGraph False) Nothing
+
+extractModuleGraphInfo :: ModuleGraph -> ModuleGraphInfo
+extractModuleGraphInfo modGraph = do
+  let (graph, _) = moduleGraphNodes False (mgModSummaries' modGraph)
+      vtxs = G.verticesG graph
+      modNameFromVertex =
+        fmap getModuleNameText . gnode2ModSummary . G.node_payload
+      modNameMapLst =
+        mapMaybe
+          (\v -> (G.node_key v,) <$> modNameFromVertex v)
+          vtxs
+      modNameMap :: IntMap ModuleName
+      modNameMap = IM.fromList modNameMapLst
+      modNameRevMap :: Map ModuleName Int
+      modNameRevMap = M.fromList $ fmap swap modNameMapLst
+      topSorted =
+        mapMaybe
+          (\n -> M.lookup n modNameRevMap)
+          $ getTopSortedModules modGraph
+      modDeps = IM.fromList $ fmap (\v -> (G.node_key v, G.node_dependencies v)) vtxs
+   in ModuleGraphInfo modNameMap modDeps topSorted
+
+getModuleNameFromPipeState :: PipeState -> Maybe ModuleName
+getModuleNameFromPipeState pstate =
+  let mmi = iface pstate
+      mmod = fmap mi_module mmi
+   in fmap (T.pack . moduleNameString . moduleName) mmod
+
+formatName :: DynFlags -> Name -> String
+formatName dflags name =
+  let str = showSDoc dflags . ppr . localiseName $ name
+   in case str of
+        (x : _) ->
+          -- NOTE: As we want to have resultant text directly copied and pasted to
+          --       the source code, the operator identifiers should be wrapped with
+          --       parentheses.
+          if isAlpha x
+            then str
+            else "(" ++ str ++ ")"
+        _ -> str
+
+formatImportedNames :: [String] -> String
+formatImportedNames names =
+  case fmap (++ ",\n") $ L.sort names of
+    l0 : ls ->
+      let l0' = "  ( " ++ l0
+          ls' = fmap ("    " ++) ls
+          footer = "  )"
+       in concat ([l0'] ++ ls' ++ [footer])
+    _ -> "  ()"
+
+mkModuleNameMap :: GlobalRdrElt -> [(ModuleName, Name)]
+mkModuleNameMap gre = do
+  spec <- gre_imp gre
+  case gre_name gre of
+    NormalGreName name -> do
+      let modName = T.pack . moduleNameString . is_mod . is_decl $ spec
+      pure (modName, name)
+    -- TODO: Handle the record field name case correctly.
+    FieldGreName _ -> []
